requirements.txt
Django==1.6.6
djangorestframework==2.4.3
requests==2.4.1
wsgiref==0.1.2
